+++ Backend/responses-code.js
@@ -1,9 +1,11 @@
 const SERVER_ERROR = 500;
 const VALIDATION_ERROR = 422;
+const AUTHENTICATION_ERROR = 411;
 const SUCCESS = 200
 
 module.exports = {
     SERVER_ERROR,
     VALIDATION_ERROR,
+    AUTHENTICATION_ERROR,
     SUCCESS
 }
